% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{add_frequency}
\alias{add_frequency}
\title{Add frequency of words or other features}
\usage{
add_frequency(data, feature = word, by = doc_id, label = NULL)
}
\arguments{
\item{data}{A tidy data frame, potentially containing a column called "word"}

\item{feature}{The feature to count in each document}

\item{by}{A grouping column identifying a document, such as \code{doc_id}}

\item{label}{Whether to label variables added to data frame}
}
\value{
The original data frame with a column added for frequency
}
\description{
Add frequency of words or other features
}
\examples{
\dontrun{
  my_corpus <- load_texts()

  my_bigrams <- my_corpus |>
    add_frequency()
}

\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  add_frequency() |>
  head()
}
}
