% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_progress}
\alias{add_progress}
\alias{add_index}
\title{Track progress in documents}
\usage{
add_progress(
  data,
  by = doc_id,
  unit = c("percentage", "index"),
  name = progress,
  label = NULL
)

add_index(data, by = doc_id, name = word_index, label = NULL)
}
\arguments{
\item{data}{A tidy data frame, divided by lines, words, or some other feature}

\item{by}{A grouping column identifying a document, such as \code{doc_id}}

\item{unit}{Unit for measuring progress, either as a percentage or as an index of row number}

\item{name}{The name to use for the added column}

\item{label}{Whether to label variables added to data frame}
}
\value{
A data frame with one additional column
}
\description{
Track progress in documents
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  add_index() |>
  add_sentiment() |>
  drop_na() |>
  head()
}
}
