% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse gt rows in the style of kableExtra}
\usage{
collapse_rows(df_g, col, lookleft = TRUE)
}
\arguments{
\item{df_g}{A gt table data object}

\item{col}{The column to collapse}

\item{lookleft}{Whether to depend any collapsing on the column one step left}
}
\value{
A gt table data object
}
\description{
Collapse gt rows in the style of kableExtra
}
\examples{
library(tibble)
library(dplyr)
library(gt)
mtcars |>
  rownames_to_column("model") |>
  relocate(cyl, gear) |>
  arrange(cyl, gear, mpg) |>
  gt() |>
  collapse_rows(cyl) |>
  collapse_rows(gear)
}
\seealso{
Other table helpers: 
\code{\link{tabulize}()}
}
\concept{table helpers}
