% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{combine_ngrams}
\alias{combine_ngrams}
\title{Combine ngram columns}
\usage{
combine_ngrams(data, feature = word, keep = FALSE, label = NULL)
}
\arguments{
\item{data}{A tidy data frame, potentially containing columns called "word_1", "word_2", etc.}

\item{feature}{The column name prefix for the numbered columns of ngrams}

\item{keep}{Whether to keep the original columns called "word_1", "word_2", etc., alongside the new "ngram" column}

\item{label}{Whether to label variables added to data frame}
}
\value{
A data frame with a column called ngram
}
\description{
Combine ngram columns
}
\examples{
\dontrun{
  my_corpus <- load_texts(n = 2)

  my_bigrams <- my_corpus |>
    add_ngrams(collapse = FALSE) |>
    combine_ngrams()

dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  add_ngrams(2) |>
  combine_ngrams() |>
  head()
}
}
\seealso{
Other n-gram helpers: 
\code{\link{add_ngrams}()},
\code{\link{plot_bigrams}()},
\code{\link{separate_ngrams}()}
}
\concept{n-gram helpers}
