% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowed.R
\name{count}
\alias{count}
\title{Count values in one or more columns}
\usage{
count(x, ...)
}
\arguments{
\item{x}{A data frame}

\item{...}{The columns to count. If more than one column is chosen, combinations of values are counted.}
}
\value{
A data frame with one column for every grouping identified by \code{...} and an additional column for the number of values counted
}
\description{
Count values in one or more columns
}
\examples{
mtcars |>
  count(cyl)
}
