% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty.R
\name{drop_empty}
\alias{drop_empty}
\alias{tidyr::drop_na}
\title{Drop rows containing missing values}
\usage{
drop_empty(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Columns to inspect for missing values. If empty, all columns are used.}
}
\description{
\code{drop_empty()} drops rows where any column specified by \code{...} contains a missing value. Internally, the function repurposes \code{\link[tidyr:drop_na]{tidyr::drop_na()}}.
}
