% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{drop_labels}
\alias{drop_labels}
\title{Drop variable labels}
\usage{
drop_labels(data, vars = NULL)
}
\arguments{
\item{data}{A data frame possibly containing labels}

\item{vars}{A vector of column names to be stripped of their labels. If this optional argument is not supplied, labels will be dropped for every column}
}
\value{
An unlabeled data frame in the same format as \code{data}
}
\description{
Drop variable labels
}
\examples{
\dontrun{
  drop_labels(vocab_dubliners)
}
}
