% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_documents.R
\name{expand_documents}
\alias{expand_documents}
\title{Convert data frame from long tidy format to wider format}
\usage{
expand_documents(
  data,
  feature = word,
  by = doc_id,
  percent = TRUE,
  sort = TRUE,
  columns = NULL
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing a column called "word"}

\item{feature}{A column of words containing one word per row, to be counted for frequency}

\item{by}{A column containing document grouping}

\item{percent}{Whether frequencies should be converted to percentages on a per-document basis}

\item{sort}{Whether to sort features by frequency}

\item{columns}{The features to keep}
}
\value{
A data frame with one row per document and as many features as words.
}
\description{
The resulting data frame is a simpler form of the document feature matrix used by other packages. \code{my_df |> expand_documents(percent = FALSE, sort = FALSE)} compares to \code{my_df |> count(doc_id, word) |> tidytext::cast_dfm(doc_id, word, n)}, but it is not equivalent. The latter prepares a DFM to be used with the quanteda package.
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  expand_documents()
}
}
