% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_texts.R
\name{get_corpus}
\alias{get_corpus}
\title{Prepare a corpus or corpora of texts}
\usage{
get_corpus(
  corpus,
  name = ".txt",
  word = TRUE,
  lemma = FALSE,
  lemma_replace = FALSE,
  to_lower = TRUE,
  remove_names = FALSE,
  pos = FALSE,
  poetry = FALSE,
  paragraph = TRUE,
  cache = TRUE
)
}
\arguments{
\item{corpus}{Vector of any length, where each value is either a string identifying a directory of texts or the first part of a filename to a cached RDS file prepared by tmtyro.}

\item{name}{What naming pattern to search for in this folder. Defaults to ".txt".}

\item{word}{Whether to split one word per line. Defaults to TRUE.}

\item{lemma}{Whether to lemmatize the text. When \code{word} is TRUE, adds a new column called \code{lemma}. This step can add a lot of time, so it defaults to FALSE.}

\item{lemma_replace}{When \code{lemma} and \code{word} are both TRUE, toggles whether to replace the \code{word} column with the lemmatized tokens. Defaults to FALSE}

\item{to_lower}{When \code{word} is TRUE, toggles whether to convert all words to lowercase. Defaults to TRUE.}

\item{remove_names}{When \code{word} is TRUE, toggles whether to remove words that only appear with the form of initial capitals. Defaults to FALSE.}

\item{pos}{Whether to add a column for part-of-speech tag. This step can add a lot of time, so it defaults to FALSE.}

\item{poetry}{Whether to detect and indicate stanza breaks and line breaks. Defaults to FALSE.}

\item{paragraph}{Whether to detect paragraph breaks for prose. Defaults to TRUE.}

\item{cache}{Whether to save a cached copy of the corpus. Some options like \code{pos = TRUE} and \code{lemma = TRUE} can add significant time to corpus preparation, so setting \code{cache = TRUE} saves the need to repeat steps each time a corpus is loaded. Defaults to TRUE.}
}
\value{
A data frame with columns for corpus, doc_id, and other data.
}
\description{
\code{get_corpus()} works nearly identically as \code{load_texts()}, but it has two fundamental differences. First, it adds a "corpus" column to the resulting table to help with record keeping. Second, it adds an option for caching its output in a local RDS file, saved in the project directory.
}
\examples{
\dontrun{
  austen <- get_corpus("austen")

  shakespeare <- get_corpus(
    c("comedy",
      "history",
      "tragedy"))
}
}
