% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_data_dictionary}
\alias{get_data_dictionary}
\title{Prepare a data dictionary from column labels}
\usage{
get_data_dictionary(data, only_labeled = FALSE)
}
\arguments{
\item{data}{A tmtyro data frame}

\item{only_labeled}{Whether to return only the variables with descriptive labels}
}
\value{
data frame with three columns: \code{variable}, \code{label}, and \code{class}
}
\description{
Prepare a data dictionary from column labels
}
\examples{
\dontrun{
  get_data_dictionary(vocab_dubliners)
}
}
