% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_df_by}
\alias{get_df_by}
\title{Get document frequencies of values in one vector \code{x} categorized by another vector \code{by}.}
\usage{
get_df_by(x, by, percent = TRUE)
}
\arguments{
\item{x}{A vector, such as a column of character strings}

\item{by}{A vector of categories, such as a column of document identifiers}

\item{percent}{Whether to return frequencies as percentage of the whole}
}
\value{
A vector of document frequencies for each value pair of \code{x} and \code{by}.
}
\description{
Get document frequencies of values in one vector \code{x} categorized by another vector \code{by}.
}
\examples{
my_values <- c(
  "the", "cat", "was", "bad",
  "the", "dog", "was", "very", "good",
  "the", "lizard", "is", "the", "most", "bad")
my_docs <- c(
  "A", "A", "A", "A",
  "B", "B", "B", "B", "B",
  "C", "C", "C", "C", "C", "C")

get_df_by(my_values, my_docs)
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_frequency}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
