% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg.R
\name{get_gutenberg_corpus}
\alias{get_gutenberg_corpus}
\title{Build and load a corpus from Project Gutenberg}
\usage{
get_gutenberg_corpus(
  gutenberg_id,
  download = c("auto", "always", "temp", "never"),
  dir = "gutenberg",
  meta_fields = c("gutenberg_id", "title", "author"),
  html_title = FALSE,
  ...
)
}
\arguments{
\item{gutenberg_id}{A vector of ID numbers from Project Gutenberg or a data frame containing a \code{gutenberg_id} column, such as from the results of a call to \code{\link[gutenbergr:gutenberg_works]{gutenbergr::gutenberg_works()}}}

\item{download}{Whether files should be automatically downloaded into a project subdirectory as needed (the default), always downloaded into the project folder, temporarily downloaded once per-session, or never downloaded}

\item{dir}{The project subdirectory for storing downloaded \code{.htm} files}

\item{meta_fields}{Additional fields to add from \link[gutenbergr:gutenberg_metadata]{gutenbergr::gutenberg_metadata} describing each book}

\item{html_title}{Whether to use the h1 header from an HTML file to determine a document's title instead of \link[gutenbergr:gutenberg_metadata]{gutenbergr::gutenberg_metadata}}

\item{...}{Additional parameters passed along to \code{\link[gutenbergr:gutenberg_strip]{gutenbergr::gutenberg_strip()}}}
}
\value{
A data frame with one row for each line of the texts in the corpus
}
\description{
\code{get_gutenberg_corpus()} improves upon the functionality
of \code{\link[gutenbergr:gutenberg_download]{gutenbergr::gutenberg_download()}} in three key ways.
\enumerate{
\item Retrieving the ".htm" version of texts
instead of the ".zip" version typically used by
gutenberger dramatically improves file coverage.
\item Parsing HTML headers allows texts to be studied
by sections and chapters. Parsing is handled by
\code{\link[=parse_html]{parse_html()}}, with \code{\link[=move_header_to_text]{move_header_to_text()}}
available for corrections.
\item Caching files locally avoids repeated
downloads, thereby improving code portability,
allowing offline access, and reducing network use.
}
All changes are made with consideration for server bandwidth,
so a two-second delay is introduced between each download
attempt. This will slow down the initial acquisition of
corpora, but offline caching speeds things up considerably
in subsequent use.
}
\examples{
\dontrun{
library(gutenbergr)

dalloway <- gutenberg_works(author == "Woolf, Virginia",
                            title == "Mrs Dalloway in Bond Street") |>
  get_gutenberg_corpus()
}
}
