% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_hir}
\alias{get_hir}
\title{Cumulative hapax introduction ratio}
\usage{
get_hir(x)
}
\arguments{
\item{x}{A vector, such as a column of character strings}
}
\value{
A vector of ratios
}
\description{
Hapax introduction ratio tracks the use of unique words as a text grows.
}
\examples{
c("cat", "dog", "dog", "bat", "dog") |>
  get_hir()
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_df_by}()},
\code{\link{get_frequency}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
