% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{interactive_document_topics}
\alias{interactive_document_topics}
\title{Explore topics interactively}
\usage{
interactive_document_topics(
  lda,
  top_n = 4,
  title = FALSE,
  height = NULL,
  omit = NULL,
  smooth = TRUE
)
}
\arguments{
\item{lda}{The topic model to be used.}

\item{top_n}{The number of topics to visualize. By default, the top 4 topics in each document will be shown.}

\item{title}{By default, the function will add a title to the chart, corresponding to the name of the object passed to the \code{lda} parameter. Set it to FALSE to return a chart with no title.}

\item{height}{The height of the resulting HTML widget.}

\item{omit}{Upon exploration, some topics may be found to contain common stop words or other unhelpful material. Use the \code{omit} parameter to define a vector of topic numbers you wish to omit from a visualization.}

\item{smooth}{After samples are rejoined, the measured value of each topic will vary wildly, even in samples that are beside each other in a document. This can make charts distractingly jittery. The default TRUE value of this parameter reduces chart noise by calculating rolling averages across three samples. Set the parameter to FALSE to skip this step and allow for visualization of extreme values.}
}
\value{
Interactive plotly object
}
\description{
\code{interactive_document_topics()} uses plotly to prepare an interactive visualization to explore a topic model, showing the top "n" topics in each document. This kind of visualization is for use in the interactive IDE or as a web page.
}
\examples{
\dontrun{
mysteries <- load_texts("mystery-novels", word = FALSE)

mysteries_lda <- mysteries |>
  make_topic_model(k = 10)

interactive_document_topics(mysteries_lda, top_n = 5)
}
}
