% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{italicize_titles}
\alias{italicize_titles}
\title{Italicize document titles in a table}
\usage{
italicize_titles(df_g, col = doc_id)
}
\arguments{
\item{df_g}{A gt table data object}

\item{col}{The column to italicize}
}
\value{
A gt table data object
}
\description{
Italicize document titles in a table
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(gt)
library(palmerpenguins)
library(tmtyro)

penguins_gt <-
  penguins |>
  select(-year) |>
  summarize(
    across(
      ends_with("_mm"), mean, na.rm = TRUE),
    .by = c(species, island, sex)) |>
  gt() |>
  fmt_number() |>
  tab_spanner(
    "bill",
    columns = starts_with("bill_")) |>
  tab_spanner(
    "flipper",
    starts_with("flip")) |>
  cols_label(
    bill_length_mm = "length",
    bill_depth_mm = "depth",
    flipper_length_mm = "length") |>
  sub_missing()

penguins_gt |>
  collapse_rows(species) |>
  collapse_rows(island) |>
  italicize_titles(species)
}\if{html}{\out{</div>}}
}

\seealso{
Other table helpers: 
\code{\link{collapse_rows}()},
\code{\link{tabulize}()}
}
\concept{table helpers}
