% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{load_topic_model}
\alias{load_topic_model}
\title{Load (or cache and load) a topic model}
\usage{
load_topic_model(
  data,
  k = 15,
  by = doc_id,
  sample_size = 1000,
  lda_name = NULL
)
}
\arguments{
\item{data}{A data frame with nested text in a "text" column.}

\item{k}{The number of topics to search for. By default, 15 topics will be sought.}

\item{by}{The column for identifying each document. By default, the "title" column will be used.}

\item{sample_size}{The sample size for each document chunk. By default, samples will include 1000 words.}

\item{lda_name}{The name to use when looking caching the model in the "data" folder. By default, this value is derived from the name of the data frame passed into the function, but it may be helpful to define it explicitly if there are intervening steps before loading the topic model.}
}
\value{
A topic model.
}
\description{
\code{load_topic_model()} checks to see whether a cached topic model exists before creating one with \code{\link[=make_topic_model]{make_topic_model()}} and caching it.
}
\examples{
\dontrun{
mysteries <- load_texts("mystery-novels", word = FALSE)

mysteries_lda <- mysteries |>
  load_topic_model(k = 10)
  }
}
