% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{make_dictionary}
\alias{make_dictionary}
\title{Create a lexicon}
\usage{
make_dictionary(list, identifier = "value")
}
\arguments{
\item{list}{A list of named word vectors}

\item{identifier}{The kind of dictionary}
}
\value{
A data frame with two columns, "word" and \if{html}{\out{<identifier>}}
}
\description{
\code{make_dictionary()} creates a dictionary for use with \code{add_dictionary()}.
}
\examples{
my_sentiments <- make_dictionary(
  list(
    "😊" = c("happy", "joy", "smile"),
    "😔" = c("unhappy", "sad", "frown")),
  identifier = "emoji")

border_states <- make_dictionary(
  list(
    "Canada" = c(
      "Alaska", "Washington", "Idaho",
      "Montana", "North_Dakota", "Minnesota",
      "Wisconsin", "Michigan", "Ohio",
      "Pennsylvania", "New_York", "Vermont",
      "New_Hampshire", "Maine", "AK", "WA",
      "ID", "MT", "ND", "MN", "WI", "MI",
      "OH", "PA", "NY", "VT", "NH", "ME"),
    "Mexico" = c(
      "California", "Arizona", "New_Mexico",
      "Texas", "CA", "AZ", "NM", "TX")),
  identifier = "borders")
}
