% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{make_topic_model}
\alias{make_topic_model}
\title{Construct a topic model}
\usage{
make_topic_model(data, by = doc_id, sample_size = 1000, k = 15, cache = TRUE)
}
\arguments{
\item{data}{A data frame with nested text in a "text" column.}

\item{by}{The column for identifying each document. By default, the "title" column will be used.}

\item{sample_size}{The sample size for each document chunk. By default, samples will include 1000 words.}

\item{k}{The number of topics to search for. By default, 15 topics will be sought.}

\item{cache}{Whether to cache the resulting model as an RDS file in the "data/" folder. Set to \code{TRUE} by default. Delete this RDS file to create a new model.}
}
\value{
A topic model.
}
\description{
\code{make_topic_model()} moves a table of texts through the necessary steps of preparation before building a topic model. The function applies seven steps:
\enumerate{
\item identifies text divisions by the \code{doc_id} column
\item divides each of the texts into same-sized chunks of \code{sample_size} words (default is 1000 words)
\item unnests text table into a table with one word per row
\item removes stop words and proper nouns (identified as any word that only appears with a capitalized first letter)
\item counts word frequencies for each chunk
\item converts the table of frequencies into a document term matrix
\item builds a topic model with \code{k} topics
}
}
\examples{
\dontrun{
mysteries <- load_texts("mystery-novels", word = FALSE)

mysteries_lda <- mysteries |>
  make_topic_model(k = 10)
  }
}
