% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{measure_tf_idf}
\alias{measure_tf_idf}
\title{Add measures for tf-idf}
\usage{
measure_tf_idf(df, by = doc_id, feature = word)
}
\arguments{
\item{df}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{by}{A column containing document grouping}

\item{feature}{A column containing the terms to be measured across document groupings}
}
\value{
A summary of the original data frame, with rows for each document and term pairing and columns for document identifier, term, n (the number of times this term was used in this document), tf (term's frequency in this document), idf (inverse document frequency), and tf_idf (previous two columns combined).
}
\description{
Add measures for tf-idf
}
\examples{
library(tmtyro)
library(dplyr)

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  measure_tf_idf()
}
