% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_micusp.R
\name{micusp_metadata}
\alias{micusp_metadata}
\title{Get MICUSP metadata}
\usage{
micusp_metadata(micusp_dir = "micusp")
}
\arguments{
\item{micusp_dir}{The name of the directory for storing local copies of MICUSP materials. Defaults to "micusp/".}
}
\value{
A data frame with 1 row for each document in the corpus and 8 columns of metadata: \code{paper_id}, \code{title}, \code{discipline}, \code{paper_type}, \code{student_level}, \code{sex}, \code{nativeness}, and \code{textual_features}
}
\description{
Explore metadata of available MICUSP (Michigan Corpus of Upper-level Student Papers) texts to choose a corpus. On first use, the function creates a folder in the working directory and downloads the "micusp_metadata.csv" file before opening it. Subsequent use of the function will load from the local copy.
}
\examples{
\dontrun{
micusp_metadata() |> head()
}
}
