% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg.R
\name{move_header_to_text}
\alias{move_header_to_text}
\title{Move a header column to text}
\usage{
move_header_to_text(data, column, ...)
}
\arguments{
\item{data}{A data frame with a column called \code{text} and at least one other column indicating parts, chapters, or sections.}

\item{column}{The header column to move}

\item{...}{(optional) Filtering condition, such as \code{title == "Ulysses"}.}
}
\value{
A data frame with the header column moved into \code{text}, conditional on \code{...}
}
\description{
In some texts, header tags of a particular level indicate typographical variance that shouldn't be confused with other section tags. \code{move_header_to_text()} provides a simple method to adjust the table.
}
\examples{
if (FALSE) {
  joyce2 <- joyce |>
    move_column_to_text(subsection, title == "Ulysses")
}

}
