% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative.R
\name{narrativize}
\alias{narrativize}
\title{Describe the steps taken}
\usage{
narrativize(
  data,
  format = c("bullets", "text"),
  dictionary = narrative_dictionary_en,
  return = c("message", "character", "print"),
  person = c(NA, "we", "I")
)
}
\arguments{
\item{data}{Data processed with one or more functions from tmtyro}

\item{format}{The format of output to return. Use \code{format = "bullets"} for well formatted bullets and \code{format = "text"} for paragraph-formatted text}

\item{dictionary}{The narrative dictionary to use. \code{\link{narrative_dictionary_en}} is defined by default.}

\item{return}{How output should be returned, whether using \code{message()}, \code{print()}, or as a standard character string}

\item{person}{The personal pronoun to use in the narrative, if any. This parameter depends on the narrative dictionary defined in \code{dictionary}.}
}
\value{
Text
}
\description{
\code{narrativize()} provides a simple method for sharing methods used to prepare an analysis.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles(drop_articles = TRUE) |>
  add_frequency()

dubliners |>
   narrativize()
}\if{html}{\out{</div>}}

Returns the following:
\itemize{
\item Retrieved a corpus from Project Gutenberg using ID number 2814.
\item Loaded texts by tokenizing words, converting to lowercase, and preserving paragraph breaks.
\item Identified documents using the \code{part} column.
\item Standardized titles by converting to title case.
\item Counted the frequency of words used in each document.
}
}

