% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg.R
\name{parse_html}
\alias{parse_html}
\title{Read HTML headers and text from file}
\usage{
parse_html(html, headers = 1:6, standardize_headers = TRUE, title = TRUE)
}
\arguments{
\item{html}{A file in HTML format}

\item{headers}{The HTML header levels to consider}

\item{standardize_headers}{Whether to standardize HTML headers to useful column names}

\item{title}{Whether to keep H1 tags even when there is only one unique value}
}
\value{
A data frame with a column called "text" and header columns limited to page elements like h1, h2, and h3, as included in the numeric range of \code{headers}
}
\description{
Read HTML headers and text from file
}
\examples{
if (FALSE) {
  library(dplyr)
  library(stringr)
  library(tmtyro)

  orlando <-
    "http://gutenberg.net.au/ebooks02/0200331h.html" |>
    download_once() |>
    parse_html() |>
    filter(str_detect(part, "CHAPTER")) |>
    mutate(
      chapter = str_extract(part, "\\\\d"),
      author = "Virginia Woolf") |>
    select(author, title, chapter, text) |>
    drop_na(chapter) |>
    identify_by(title, chapter) |>
    load_texts()
}

}
