% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{plot_bigrams}
\alias{plot_bigrams}
\title{Visualize bigram chains}
\usage{
plot_bigrams(
  data,
  feature = word,
  random_seed = TRUE,
  set_seed = NULL,
  legend = FALSE,
  top_n = 35
)
}
\arguments{
\item{data}{A tidy data frame potentially containing a column called "word" or columns called "word_1" and "word_2".}

\item{feature}{The feature to use when constructing ngrams}

\item{random_seed}{Whether to randomize the creation of the network chart.}

\item{set_seed}{A specific seed to use if not random}

\item{legend}{Whether to show a legend for the edge color}

\item{top_n}{The number of pairs to visualize}
}
\value{
A ggplot2 object
}
\description{
Visualize bigram chains
}
\examples{
\dontrun{
# It isn't necessary to use add_ngrams()
data |>
  plot_bigrams()

# Adding them first allows for filtering steps
data |>
  add_ngrams() |>
  drop_stopwords(word_1) |>
  drop_stopwords(word_2) |>
  plot_bigrams()

# Only bigrams influence the visualization These show the same networks:
data |>
  add_ngrams() |>
  plot_bigrams()

data |>
  add_ngrams(4) |>
  plot_bigrams()

dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  plot_bigrams()

# Loading `ggraph` enables edge to show connection strengths
library(ggraph)

dubliners |>
  plot_bigrams()

dubliners |>
  add_ngrams(2) |>
  drop_stopwords(feature = word_1) |>
  drop_stopwords(feature = word_2) |>
  plot_bigrams()

dubliners |>
  dplyr::filter(doc_id == "The Dead") |>
  plot_bigrams(top_n = 70) |>
  change_colors(c("black", "orange"))
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{italicize_titles}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}

Other n-gram helpers: 
\code{\link{add_ngrams}()},
\code{\link{combine_ngrams}()},
\code{\link{separate_ngrams}()}
}
\concept{n-gram helpers}
\concept{visualizing helpers}
\keyword{internal}
