% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_document_top_n.R
\name{plot_doc_word_heatmap}
\alias{plot_doc_word_heatmap}
\title{Plot a heatmap of ranked features}
\usage{
plot_doc_word_heatmap(
  data,
  rows = 1:10,
  by = doc_id,
  feature = word,
  label = TRUE
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{rows}{The ranks to show, not counting ties}

\item{by}{The column used for document grouping, with doc_id as the default}

\item{feature}{The column to measure, as in "word" or "lemma"}

\item{label}{Whether to show the rank as a label in the heatmap}
}
\value{
A ggplot object
}
\description{
Plot a heatmap of ranked features
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts(lemma = TRUE) |>
  identify_by(part) |>
  standardize_titles()

# Make a smaller example
selected_titles <-
  c("The Sisters", "An Encounter", "Araby",
    "Counterparts", "The Dead")

dubliners |>
  dplyr::filter(doc_id \%in\% selected_titles) |>
  plot_doc_word_heatmap()

dubliners |>
  dplyr::filter(doc_id \%in\% selected_titles) |>
  plot_doc_word_heatmap(feature = lemma, rows = 1:6)
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{italicize_titles}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
\keyword{internal}
