% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_variety.R
\name{plot_hapax}
\alias{plot_hapax}
\title{Project hapax legomena onto vocabulary growth}
\usage{
plot_hapax(
  data,
  prop = 0.01,
  x = progress,
  y = vocabulary,
  by = doc_id,
  descriptive_labels = TRUE,
  feature = hapax_doc
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{prop}{The proportion of hapax to sample. The chart can become illegible with proportions over ~1\%}

\item{x}{A column showing the cumulative progress of documents}

\item{y}{The Y-axis variable to chart. Default value is the cumulative vocabulary size.}

\item{by}{A grouping column, such as doc_id}

\item{descriptive_labels}{A toggle for disabling descriptive labels of progress_percent on the X-axis}

\item{feature}{The column to check for new features. Defaults to \code{hapax_doc}, but the function might also be used with \code{new_word} instead to plot a sample of new additions to documents' vocabularies.}
}
\value{
A ggplot object
}
\description{
\code{plot_hapax()} visualizes a sampling of hapax legomena projected on faceted curves of vocabulary growth over time
}
\examples{
if (FALSE) {
  dubliners <- get_gutenberg_corpus(2814) |>
    load_texts() |>
    identify_by(part) |>
    standardize_titles()

  dubliners_measured <- dubliners |>
    add_vocabulary()

  dubliners_measured |>
    plot_hapax()
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{italicize_titles}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
\keyword{internal}
