% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{plot_ngrams}
\alias{plot_ngrams}
\title{Visualize ngram chains}
\usage{
plot_ngrams(
  df,
  feature = word,
  random_seed = TRUE,
  set_seed = NULL,
  node_color = "lightblue",
  edge_color = NULL,
  top_n = 25
)
}
\arguments{
\item{df}{A tidy data frame potentially containing a column called "word" or columns called "word_1" and "word_2".}

\item{feature}{The feature to use when constructing ngrams}

\item{random_seed}{Whether to randomize the creation of the network chart.}

\item{set_seed}{A specific seed to use if not random}

\item{node_color}{The color of the nodes}

\item{edge_color}{The color of the edges. When the \code{ggraph} package is not loaded, the default \code{edge_color=NULL} setting presents edges as gray. When that package is is loaded, a \code{NULL} edge color colors edges black. Whatever the color defined by this parameter, loading \code{ggraph} enables transparency to show strength of each bigram connection.}

\item{top_n}{The number of pairs to visualize}
}
\value{
A ggplot2 object
}
\description{
Mildly adapted from \href{https://www.tidytextmining.com}{Text Mining with R}, by Julia Silge and David Robinson
}
\examples{
\dontrun{
# It isn't necessary to use add_ngrams()
df |>
  plot_ngrams()

# Adding them first allows for filtering steps
df |>
  add_ngrams() |>
  drop_stopwords(word_1) |>
  drop_stopwords(word_2) |>
  plot_ngrams()

# Only bigrams influence the visualization These show the same networks:
df |>
  add_ngrams(2) |>
  plot_ngrams()

df |>
  add_ngrams(4) |>
  plot_ngrams()

}

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  plot_ngrams()

library(ggraph)

austen |>
  plot_ngrams()

austen |>
  add_ngrams(2) |>
  drop_stopwords(feature = word_1) |>
  drop_stopwords(feature = word_2) |>
  plot_ngrams()

austen |>
  dplyr::filter(doc_id == "northanger") |>
  plot_ngrams(top_n = 90, node_color = "lightgreen")
}
