% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{plot_topic_bars}
\alias{plot_topic_bars}
\title{Plot bars for words in each topic}
\usage{
plot_topic_bars(
  lda,
  topics,
  top_n = 10,
  expand_bars = TRUE,
  save = TRUE,
  saveas = "png",
  savedir = "plots"
)
}
\arguments{
\item{lda}{The topic model to be used.}

\item{topics}{The topic numbers to view}

\item{top_n}{The number of words to show for each topic}

\item{expand_bars}{Whether to stretch the bars the length of the X-axis for each facet}

\item{save}{By default, the visualization will be saved. Set to FALSE to skip saving.}

\item{saveas}{The filetype for saving resulting visualizations. By default, the files will be in "png" format, but other options such as "pdf" or "jpg will also work.}

\item{savedir}{The directory for saving output images. By default, this is set to "plots/".}
}
\value{
A ggplot2 visualization showing the top words in each of the chosen topics.
}
\description{
Plot bars for words in each topic
}
\examples{
\dontrun{
austen <-
  get_gutenberg_corpus(c(105, 121, 141, 158, 161, 946, 1342)) |>
  select(doc_id = title, text)

# Creating the topic model for the package
austen_lda <- make_topic_model(austen, k = 30)

saveRDS(austen_lda, "inst/austen_lda.rds")
}

# Loading from the package
austen_lda <- "austen_lda.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen_lda |>
  plot_topic_bars(topics = c(6, 15, 19))
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_htr}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
