% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{plot_topic_wordcloud}
\alias{plot_topic_wordcloud}
\title{Plot topic wordclouds}
\usage{
plot_topic_wordcloud(df, topics = NULL, crop = TRUE, savedir = "plots")
}
\arguments{
\item{df}{A loaded topic model}

\item{topics}{Topic numbers to be visualized. If left undefined, all topics will be visualized}

\item{crop}{Whether to remove white space from visualized word clouds}

\item{savedir}{The directory to save plots in. Defaults to "plots"}
}
\value{
Graphic(s) prepared with \code{knitr} for Quarto or RMarkdown
}
\description{
\code{plot_topic_wordcloud()} prepares, saves, and displays word clouds of topics in a topic model. The function can display word clouds of one or more specific topics, or it can show word clouds for every topic.
}
\examples{
\dontrun{
mysteries <- load_texts("mystery-novels", word = FALSE)

mysteries_lda <- mysteries |>
  make_topic_model(k = 10)
  }
}
