% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos_tags}
\alias{pos_tags}
\title{Part of speech tags}
\format{
\subsection{"pos_tags"}{

A data frame with 36 rows and 2 columns
\describe{
\item{tag}{speech tag}
\item{description}{meaning of tag}
}
}
}
\source{
\url{https://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}
}
\usage{
pos_tags
}
\description{
Tags and descriptions for English parts of speech, designated by the Penn Treebank Project.
}
\keyword{datasets}
