% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{set_data_dictionary}
\alias{set_data_dictionary}
\title{Assign column labels from a data dictionary}
\usage{
set_data_dictionary(data, dictionary)
}
\arguments{
\item{data}{A tmtyro data frame}

\item{dictionary}{data frame containing at least two columns, \code{variable} and \code{label}}
}
\value{
labeled data frame with the same structure as \code{data}
}
\description{
Assign column labels from a data dictionary
}
\examples{
\dontrun{
  dubliners_dictionary <- data.frame(
    variable = "doc_id",
     label = "story title"
  )

  vocab_dubliners <- vocab_dubliners |>
    set_data_dictionary(dubliners_dictionary)
}
}
