% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg.R
\name{standardize_headers}
\alias{standardize_headers}
\title{Standardize column names from HTML}
\usage{
standardize_headers(data, ..., title = TRUE)
}
\arguments{
\item{data}{A tidy data frame, potentially containing header columns "h1" through "h6"}

\item{...}{Optionally, a named list of columns and values to which they should be renamed, with defaults as faullback}

\item{title}{Whether any "h1" column should be renamed to "title"}
}
\value{
A data frame with column names adjusted
}
\description{
needed tests: unnamed arguments, values repurposed from defaults, mix of named and unnamed arguments, title = FALSE/TRUE
}
\examples{
if (FALSE) {
  joyce2 <- joyce |>
    standardize_titles() |>
    move_column_to_text(subsection, title == "Ulysses")
}

}
