% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.combined_ngrams}
\alias{tabulize.combined_ngrams}
\title{Prepare a table of n-gram frequencies}
\usage{
\method{tabulize}{combined_ngrams}(data, rows = NULL, count = TRUE, digits = 2, ...)
}
\arguments{
\item{data}{data processed with one or more functions from \code{tmtyro}}

\item{rows}{Chooses rows to be shown}

\item{count}{Determines whether frequencies will be counted for individual features}

\item{digits}{The number of digits to show past the decimal point}

\item{...}{optional parameters passed along to methods}
}
\description{
Prepare a table of n-gram frequencies
}
\keyword{internal}
