% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.default}
\alias{tabulize.default}
\title{Prepare a default table view}
\usage{
\method{tabulize}{default}(data, summary = TRUE, inorder = TRUE, count = FALSE, rows = NULL, ...)
}
\arguments{
\item{data}{data processed with one or more functions from \code{tmtyro}}

\item{summary}{Indicates whether to prepare a summary table or the rows as they exist}

\item{inorder}{Indicates whether labels in the \code{doc_id} column should have their order preserved}

\item{count}{Determines whether frequencies will be counted for individual features}

\item{rows}{Chooses rows to be shown}

\item{...}{optional parameters (unused)}
}
\description{
Prepare a default table view
}
\keyword{internal}
