% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.expanded}
\alias{tabulize.expanded}
\title{Prepare a table showing a document-feature matrix}
\usage{
\method{tabulize}{expanded}(data, columns = NULL, digits = 2, ...)
}
\arguments{
\item{data}{data processed with one or more functions from \code{tmtyro}}

\item{columns}{Chooses columns to be shown}

\item{digits}{The number of digits to show past the decimal point}

\item{...}{optional parameters passed along to methods}
}
\description{
Prepare a table showing a document-feature matrix
}
\keyword{internal}
