% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize}
\alias{tabulize}
\title{Prepare a table of data}
\usage{
tabulize(.data, ...)
}
\arguments{
\item{.data}{data processed with one or more functions from \code{tmtyro}}

\item{...}{
  Arguments passed on to \code{\link[=tabulize.default]{tabulize.default}}
  \describe{
    \item{\code{summary}}{Indicates whether to prepare a summary table or the rows as they exist}
    \item{\code{inorder}}{Indicates whether labels in the \code{doc_id} column should have their order preserved}
    \item{\code{count}}{Determines whether frequencies will be counted for individual features}
    \item{\code{rows}}{Chooses rows to be shown}
  }}
}
\value{
A gt table data object
}
\description{
\code{tabulize()} provides a simple method for sharing results. Based on previous functions used, \code{tabulize()} will choose a method, resolving to one of a set of tables.
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

# A data frame with `doc_id` and `word` columns will show word counts by default
austen |>
   tabulize()

# Specifying `count = TRUE` shows word frequencies
austen |>
   tabulize(count = TRUE)

# Defining a vector for `rows` changes which are shown
austen |>
   tabulize(count = TRUE, rows = 1:3)

# Applying `tmtyro` functions prepares some other table

austen |>
   add_ngrams() |>
   tabulize()

austen |>
   summarize_tf_idf() |>
   tabulize()

austen |>
   add_vocabulary() |>
   tabulize()

if (FALSE) { # sentiment requires interaction on first load
  austen |>
     add_sentiment() |>
     tabulize()
}
}
\seealso{
Other table helpers: 
\code{\link{collapse_rows}()}
}
\concept{table helpers}
