% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.sentiment}
\alias{tabulize.sentiment}
\title{Prepare a table for sentiment analysis}
\usage{
\method{tabulize}{sentiment}(
  data,
  inorder = TRUE,
  digits = 2,
  drop_na = FALSE,
  ignore = NULL,
  rows = NULL,
  count = TRUE,
  ...
)
}
\arguments{
\item{data}{data processed with one or more functions from \code{tmtyro}}

\item{inorder}{Indicates whether labels in the \code{doc_id} column should have their order preserved}

\item{digits}{The number of digits to show past the decimal point}

\item{drop_na}{Removes rows lacking sentiment}

\item{ignore}{Removes rows matching set sentiments}

\item{rows}{Chooses rows to be shown}

\item{count}{Determines whether frequencies will be counted for sentiments}

\item{...}{optional parameters passed along to methods}
}
\description{
Prepare a table for sentiment analysis
}
\keyword{internal}
