% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.vocabulary}
\alias{tabulize.vocabulary}
\title{Prepare a table of lexical variety}
\usage{
\method{tabulize}{vocabulary}(data, digits = 3, ...)
}
\arguments{
\item{data}{data processed with one or more functions from \code{tmtyro}}

\item{digits}{The number of digits to show past the decimal point}

\item{...}{optional parameters passed along to methods}
}
\description{
Prepare a table of lexical variety
}
\keyword{internal}
