% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{theme_tmtyro}
\alias{theme_tmtyro}
\title{Apply a smart tmtyro theme}
\usage{
theme_tmtyro(
  minor = FALSE,
  grid_x = NULL,
  grid_y = NULL,
  base_theme = ggplot2::theme_minimal,
  ...
)
}
\arguments{
\item{minor}{Whether to show minor grid lines}

\item{grid_x}{Whether to show grid lines on the X-axis. When \code{NULL}, this will be determined by the data being visualized.}

\item{grid_y}{Whether to show grid lines on the Y-axis. When \code{NULL}, this will be determined by the data being visualized.}

\item{base_theme}{The base ggplot2 theme used as a starting point}

\item{...}{Additional arguments passed to base theme}
}
\value{
a ggplot2 object
}
\description{
Apply a smart tmtyro theme
}
\examples{
if (FALSE) {
library(ggplot2)

mtcars |>
  ggplot(aes(disp, hp)) +
  geom_point() +
  theme_tmtyro()

mtcars |>
  ggplot(aes(hp, factor(cyl))) +
  geom_point() +
  theme_tmtyro()

mtcars |>
  ggplot(aes(factor(cyl), hp)) +
  geom_point() +
  theme_tmtyro()
}

}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
