% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_poetry_texts.R
\name{tidy_poetry_texts}
\alias{tidy_poetry_texts}
\title{Prepare a folder of poetry}
\usage{
tidy_poetry_texts(folder = "data", name = ".txt", word = TRUE)
}
\arguments{
\item{folder}{A directory containing prose texts. Defaults to "data".}

\item{name}{What naming pattern to search for in this folder. Defaults to ".txt".}

\item{word}{Whether to split one word per line. Defaults to TRUE.}
}
\value{
A data frame.
}
\description{
Prepare a folder of poetry
}
\examples{
\dontrun{
dickinson <- tidy_poetry_texts("~/Documents/Dickinson-corpus")
}
}
