% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_prose_texts.R
\name{tidy_prose_texts}
\alias{tidy_prose_texts}
\title{Prepare a folder of prose texts}
\usage{
tidy_prose_texts(folder = "data", name = ".txt", word = TRUE, paragraph = TRUE)
}
\arguments{
\item{folder}{A directory containing prose texts. Defaults to "data".}

\item{name}{What naming pattern to search for in this folder. Defaults to ".txt".}

\item{word}{Whether to split one word per line. Defaults to TRUE.}

\item{paragraph}{Whether to detect paragraph breaks.}
}
\value{
A data frame.
}
\description{
Prepare a folder of prose texts
}
\examples{
\dontrun{
mysteries <- tidy_prose_texts("mystery-novels")
}
}
