% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmtyro-package.R
\docType{package}
\name{tmtyro-package}
\alias{tmtyro}
\alias{tmtyro-package}
\title{tmtyro: tmtyro: Simplified Workflows for Text Mining Tyros}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Work with and analyze text for simple and complex features. Adopting tidytext principles, tmtyro abstracts processes a few levels further to allow beginners to use and to familiarize themselves with techniques before they understand much code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmclawson/tmtyro}
  \item \url{https://jmclawson.github.io/tmtyro/}
  \item Report bugs at \url{https://github.com/jmclawson/tmtyro/issues}
}

}
\author{
\strong{Maintainer}: James Clawson \email{clawson@gmail.com} (\href{https://orcid.org/0009-0008-6756-0607}{ORCID})

}
\keyword{internal}
