% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{add_dictionary}
\alias{add_dictionary}
\title{Add values from a dictionary}
\usage{
add_dictionary(df, dictionary, feature = word, keep_term = NULL)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{dictionary}{A data frame with two or more columns, potentially made with \code{make_dictionary()}}

\item{feature}{The column (like "word") to use for looking up values in the dictionary}

\item{keep_term}{Whether to retain the original term value. This option is especially useful with dictionaries containing terms longer than one word in length; the NULL value will keep the term for these dictionaries while discarding it for those with terms of only one word.}
}
\value{
The original data frame with one or more columns added.
}
\description{
Add values from a dictionary
}
\examples{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

emoji_weather <- make_dictionary(
  list(
    "ï¸â˜”ï¸" = c("rain", "rains", "rainy", "raining"),
    "ï¸â›ˆï¸" = c("storm", "storms", "stormy", "storming"),
    "â˜ï¸" = c("cloud", "clouds", "cloudy"),
    "ðŸŒž" = c("sun", "sunny"),
    "ðŸŒ«ï¸" = c("fog", "fogs", "foggy", "mist", "misty"),
    "ðŸŒ¬ï¸" = c("wind", "winds", "windy"),
    "ï¸â„ï¸" = c("snow", "snows", "snowing")),
  name = "weather")

dubliners |>
   add_dictionary(emoji_weather) |>
   drop_na() |>
   head()
}
