% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{add_dictionary}
\alias{add_dictionary}
\title{Add values from a dictionary}
\usage{
add_dictionary(df, dictionary, feature = word)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{dictionary}{A data frame with two or more columns, potentially made with \code{make_dictionary()}}

\item{feature}{The column (like "word") to use for looking up values in the dictionary}
}
\value{
The original data frame with one or more columns added.
}
\description{
Add values from a dictionary
}
\examples{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

my_sentiments <- make_dictionary(
  list(
    "😊" = c("happy", "joy", "smile"),
    "😔" = c("unhappy", "sad", "frown")),
  name = "emoji")

dubliners |>
   add_dictionary(my_sentiments) |>
   drop_na() |>
   head()
}
