% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_index.R
\name{add_index}
\alias{add_index}
\title{Index document row numbers}
\usage{
add_index(df, by = doc_id, name = word_index)
}
\arguments{
\item{df}{A tidy data frame, divided by lines, words, or some other feature}

\item{by}{A grouping column identifying a document, such as \code{doc_id}}

\item{name}{The name to use for the added column}
}
\value{
A data frame with one additional column
}
\description{
Index document row numbers
}
\examples{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  add_index() |>
  add_sentiment() |>
  drop_na() |>
  head()
}
