% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{add_ngrams}
\alias{add_ngrams}
\title{Add ngram columns}
\usage{
add_ngrams(
  df,
  n = 1:2,
  feature = word,
  keep = FALSE,
  collapse = FALSE,
  by = doc_id
)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{n}{A range defining the extent of an ngram---for instance, from word 1 to word 3. Alternatively, a single number will signal the number of words to include in each ngram. Default value of \code{1:2} will produce bigrams.}

\item{feature}{The feature to use when constructing ngrams}

\item{keep}{Whether to keep the original feature column}

\item{collapse}{Whether to join the ngram parts into a single column called "ngram"}

\item{by}{A grouping column identifying a document, such as \code{doc_id}.}
}
\value{
The original data frame with columns added for subsequent parts of ngrams
}
\description{
Add ngram columns
}
\examples{
\dontrun{
  my_corpus <- load_texts()

  my_bigrams <- my_corpus |>
    add_ngrams(3)
}

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  add_ngrams(2) |>
  head()
}
