% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{add_partitions}
\alias{add_partitions}
\title{Divide documents in equal lengths}
\usage{
add_partitions(df, length = 1000, by = doc_id)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{length}{Size of each partition}

\item{by}{A column containing document grouping}
}
\value{
The original data frame with a column added for partition.
}
\description{
Divide documents in equal lengths
}
\examples{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  add_partitions() |>
  head()
}
