% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{add_sentiment}
\alias{add_sentiment}
\title{Measure sentiment}
\usage{
add_sentiment(
  df,
  lexicon = c("bing", "afinn", "loughran", "nrc", "nrc_eil", "nrc_vad"),
  word = word
)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{lexicon}{The sentiment lexicon to use from the \link{tidytext} package. Options include "bing", "afinn", "loughran", "nrc", "nrc_eil", or "nrc_vad".}

\item{word}{A column of words containing one word per row, to be used for dictionary look-up.}
}
\value{
The original data frame with one or more sentiment columns added.
}
\description{
\code{\link[=add_sentiment]{add_sentiment()}} provides simple lexicon-based measures of sentiment, comparing words in a text to one of a number of controlled dictionaries.
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
   add_sentiment("nrc") |>
   tidyr::drop_na() |>
   head()
}
