% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{change_colors}
\alias{change_colors}
\title{Choose other colors}
\usage{
change_colors(
  x,
  colorset = "brewer",
  palette = 2,
  kind = "qualitative",
  direction = 1,
  start = 1
)
}
\arguments{
\item{x}{A visualization made with \code{\link[=visualize]{visualize()}}}

\item{colorset}{Either "brewer", "viridis", "okabe-ito" or a vector of colors.}

\item{palette}{The number or name of palette (dependent on setting \code{colorset} to either "brewer" or "viridis")}

\item{kind}{Used only for Brewer palettes to match numbered palette with the specific subset}

\item{direction}{The direction colors should be applied to the data. Setting to anything other than 1 will reverse the order.}

\item{start}{Useful for predefined colorsets from Okabe-Ito and Brewer to start from a color other than 1}
}
\value{
A ggplot2 object
}
\description{
\code{change_colors()} standardizes three methods for choosing color palettes for color or fill mapping, providing access to Brewer and Viridis palettes alongside custom choices.
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

# By default, Brewer's "Dark2" palette is applied to categorical data
austen |>
  visualize() |>
  change_colors()

austen |>
  add_lexical_variety() |>
  visualize(labeling = "axis") |>
  change_colors()

# Brewer's "BuGn" palette is applied to sequential data by default
austen |>
  visualize(type = "heatmap") |>
  change_colors()

austen |>
  visualize() |>
  change_colors("viridis")

# Palettes can be numbered or named
austen |>
  visualize(type = "heatmap") |>
  change_colors("viridis", palette = 6)

austen |>
  visualize(type = "heatmap") |>
  change_colors("viridis", palette = "mako")

# Manually listing colors will adopt them. When
# the number of colors differs from what's needed,
# other colors will be filled in between.
austen |>
  visualize() |>
  change_colors(c("brown", "red", "orange", "yellow", "#aaff00"))
}
\seealso{
Other visualizing helpers: 
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_htr}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
