% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{combine_ngrams}
\alias{combine_ngrams}
\title{Combine ngram columns}
\usage{
combine_ngrams(df, feature = word, keep = FALSE)
}
\arguments{
\item{df}{A tidy data frame, potentially containing columns called "word_1", "word_2", etc.}

\item{feature}{The column name prefix for the numbered columns of ngrams}

\item{keep}{Whether to keep the original columns called "word_1", "word_2", etc., alongside the new "ngram" column.}
}
\value{
A data frame with a column called ngram
}
\description{
Combine ngram columns
}
\examples{
\dontrun{
  my_corpus <- load_texts(n = 2)

  my_bigrams <- my_corpus |>
    add_ngrams(collapse = FALSE) |>
    combine_ngrams()
}

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  add_ngrams(2) |>
  combine_ngrams() |>
  head()
}
