% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize.R
\name{contextualize}
\alias{contextualize}
\title{Show a term in context}
\usage{
contextualize(
  df,
  term,
  window = 3,
  limit = 1:5,
  by = doc_id,
  feature = NULL,
  match = word,
  regex = NULL,
  html = NULL
)
}
\arguments{
\item{df}{A data frame which most likely contains a column called "word"}

\item{term}{The term to search for, exactly}

\item{window}{The number of terms to show before and after}

\item{limit}{The number of results to return in the console using cli, if installed}

\item{by}{The document identifier, for limiting context window}

\item{feature}{The column to show for context. When \code{NULL}, \code{contextualize()} looks first for an "original" column and then for a "word" column.}

\item{match}{The column to use for matching}

\item{regex}{When defined, a regular expression for searches using greater control}

\item{html}{Force message output to HTML format (e.g., for Shiny)}
}
\value{
Invisibly, a data frame with four columns: \verb{<by>}, \verb{<match>}, "index", and "context"
}
\description{
Show a term in context
}
\section{Hiding results}{
\code{contextualize()} uses \code{\link[cli:ansi-styles]{cli::style_underline()}} and \code{\link[fansi:to_html]{fansi::to_html()}}, if these packages are installed, to show formatted results in the console or in a document rendered to HTML. These formatted results can be hidden by setting \code{limit = 0} in the function, by suppressing messages with \code{\link[=suppressMessages]{suppressMessages()}} in the console, or by setting a \code{message: false} chunk option in Quarto or R Markdown.
}

\examples{
dubliners <- get_gutenberg_corpus(2814) |>
    load_texts(keep_original = TRUE)

contextualize(dubliners, regex = "dog[s]?$")
}
