% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize.R
\name{contextualize}
\alias{contextualize}
\title{Show a term in context}
\usage{
contextualize(
  df,
  term,
  window = 3,
  limit = 1:5,
  by = doc_id,
  feature = NULL,
  match = word,
  regex = NULL
)
}
\arguments{
\item{df}{A data frame which most likely contains a column called "word"}

\item{term}{The term to search for, exactly}

\item{window}{The number of terms to show before and after}

\item{limit}{The number of results to return in the console using cli, if installed}

\item{by}{The document identifier, for limiting context window}

\item{feature}{The column to show for context}

\item{match}{The column to use for matching}

\item{regex}{When defined, a regular expression for searches using greater control}
}
\value{
A data frame with four columns: \if{html}{\out{<by>}}, \if{html}{\out{<match>}}, "index", and "context"
}
\description{
Show a term in context
}
\examples{
if (FALSE) {
  dubliners <- get_gutenberg_corpus(2814) |>
    load_texts(keep_original = TRUE)

  contextualize(dubliners, regex = "dog[s]?$")
}
}
