% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty.R
\name{drop_empty}
\alias{drop_empty}
\alias{tidyr::drop_na}
\alias{drop_na}
\title{Drop rows containing missing values}
\usage{
drop_empty(data, ...)

drop_na(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Columns to inspect for missing values. If empty, all columns are used.}
}
\description{
Leave \code{...} blank to drop rows containing any missing values, or specify column names to filter specific variables. \code{drop_empty()} repurposes \code{\link[tidyr:drop_na]{tidyr::drop_na()}}.
}
