% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_stopwords.R
\name{drop_stopwords}
\alias{drop_stopwords}
\title{Remove stopwords}
\usage{
drop_stopwords(df, wordlist = NULL, feature = word)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{wordlist}{A list of stopwords}

\item{feature}{A column of words containing one word per row to be checked for stopwords.}
}
\value{
The original data frame with fewer rows.
}
\description{
Remove stopwords
}
\examples{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
   drop_stopwords()
}
