% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_stopwords.R
\name{drop_stopwords}
\alias{drop_stopwords}
\title{Remove stopwords}
\usage{
drop_stopwords(df, wordlist = stopwords::stopwords(), word = word)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "word"}

\item{wordlist}{A list of stopwords}

\item{word}{A column of words containing one word per row to be checked for stopwords.}
}
\value{
The original data frame with fewer rows.
}
\description{
Remove stopwords
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
   drop_stopwords()
}
