% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_cumulative_vocabulary}
\alias{get_cumulative_vocabulary}
\title{Cumulative total of vocabulary size}
\usage{
get_cumulative_vocabulary(x)
}
\arguments{
\item{x}{A vector, such as a column of character strings}
}
\value{
A vector of counts
}
\description{
Cumulative total of vocabulary size
}
\examples{
c("cat", "dog", "dog", "bat", "dog") |>
  get_cumulative_vocabulary()
}
\seealso{
Other vectorized functions: 
\code{\link{get_frequency}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
