% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_frequency}
\alias{get_frequency}
\alias{get_tf}
\title{Get get_frequencies of values in a vector}
\usage{
get_frequency(x, percent = FALSE)

get_tf(x, percent = TRUE)
}
\arguments{
\item{x}{A vector, such as a column of character strings}

\item{percent}{Whether to return frequencies as percentage of the whole}
}
\value{
A vector of counts or ratios for each value of \code{x}.
}
\description{
Get get_frequencies of values in a vector
}
\examples{
my_values <- c("dog", "cat", "dog")

get_frequency(my_values)
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}

Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
