% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg.R
\name{get_gutenberg_corpus}
\alias{get_gutenberg_corpus}
\title{Build and load a corpus from Project Gutenberg}
\usage{
get_gutenberg_corpus(
  gutenberg_id,
  dir = "gutenberg",
  mirror = NULL,
  meta_fields = c("title", "author"),
  ...
)
}
\arguments{
\item{gutenberg_id}{A vector of ID numbers from Project Gutenberg or a data frame containing a \code{gutenberg_id} column, such as from the results of a call to \code{\link[gutenbergr:gutenberg_works]{gutenbergr::gutenberg_works()}}.}

\item{dir}{The directory for storing downloaded \code{.txt} files. Default value is "gutenberg".}

\item{mirror}{Optionally a mirror URL to retrieve the books from. By default uses the mirror from \code{\link[gutenbergr:gutenberg_get_mirror]{gutenbergr::gutenberg_get_mirror()}}.}

\item{meta_fields}{Additional fields to add from \link[gutenbergr:gutenberg_metadata]{gutenbergr::gutenberg_metadata} describing each book. By default, title and author are added.}

\item{...}{Additional parameters passed along to \code{\link[gutenbergr:gutenberg_download]{gutenbergr::gutenberg_download()}} in the final step.}
}
\value{
A data frame with one row for each line of the texts in the corpus.
}
\description{
\code{get_gutenberg_corpus()} improves upon the functionality
of \code{\link[gutenbergr:gutenberg_download]{gutenbergr::gutenberg_download()}} in two key ways. First,
it downloads the ".txt" format of ebooks rather than the
".zip" format. This change improves coverage dramatically.
This change was made with consideration for server bandwidth,
so a two-second delay is introduced between each download
attempt. Second, it caches a copy of the file locally to avoid
repeated downloads. This change helps with code portability and
offline access.
}
\examples{
library(gutenbergr)

dalloway <- gutenberg_works(author == "Woolf, Virginia",
                            title == "Mrs. Dalloway") |>
  get_gutenberg_corpus()
}
