% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_htr}
\alias{get_htr}
\title{Cumulative hapax-token ratio}
\usage{
get_htr(x)
}
\arguments{
\item{x}{A vector, such as a column of character strings}
}
\value{
A vector of ratios
}
\description{
HTR reports the cumulative ratio of hapax legomena (one-off words) to the cumulative size of a text. This operation can be slow.
}
\examples{
c("cat", "dog", "dog", "bat", "dog") |>
  get_htr()
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_frequency}()},
\code{\link{get_hir}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
