% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_match}
\alias{get_match}
\title{Get dictionary matches of values in a vector}
\usage{
get_match(x, dictionary, keep = NULL)
}
\arguments{
\item{x}{A vector, such as a column of character strings}

\item{dictionary}{A data frame with two or more columns, potentially made with \code{make_dictionary()}.}

\item{keep}{The number of matched values to keep. If NULL, returns all matched values in a nested list.}
}
\value{
A vector or nested list of sentiments for each value of \code{x}.
}
\description{
Get dictionary matches of values in a vector
}
\examples{
my_values <- c("It", "is", "raining")

emoji_weather <- make_dictionary(
  list(
    "ï¸â˜”ï¸" = c("rain", "rains", "rainy", "raining"),
    "ï¸â›ˆï¸" = c("storm", "storms", "stormy", "storming"),
    "â˜ï¸" = c("cloud", "clouds", "cloudy"),
    "ðŸŒž" = c("sun", "sunny"),
    "ðŸŒ«ï¸" = c("fog", "fogs", "foggy", "mist", "misty"),
    "ðŸŒ¬ï¸" = c("wind", "winds", "windy"),
    "ï¸â„ï¸" = c("snow", "snows", "snowing")),
  name = "weather")

get_match(my_values, emoji_weather)
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_frequency}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
