% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_micusp.R
\name{get_micusp_corpus}
\alias{get_micusp_corpus}
\title{Prepare a corpus of texts from the Michigan Corpus of Upper-level Student Papers.}
\usage{
get_micusp_corpus(...)
}
\arguments{
\item{...}{A filter on rows and columns from \code{micusp_metadata()}. Accepted columns include the following: \code{paper_id}, \code{title}, \code{discipline}, \code{paper_type}, \code{student_level}, \code{sex}, \code{nativeness}, and \code{textual_features}.}
}
\value{
A data frame with 1 row for each document in the corpus and 9 columns. The first 8 columns contain metadata, and the final column called \code{text} contains the full text of each document.
}
\description{
The function accepts filters on columns from \code{micusp_metadata()} and downloads and parses texts locally if copies don't yet exist. It returns a table combining metadata and text data for further processing.
}
\examples{
\dontrun{
physics_f <- get_micusp_corpus(discipline == "Physics", sex == "Female")
physics_m <- get_micusp_corpus(discipline == "Physics", sex == "Female")
}
}
