% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Get sentiment matches of values in a vector}
\usage{
get_sentiment(
  x,
  lexicon = c("bing", "afinn", "loughran", "nrc", "nrc_eil", "nrc_vad"),
  ...
)
}
\arguments{
\item{x}{A vector, such as a column of character strings}

\item{lexicon}{The sentiment lexicon to use from the \link{tidytext} package. Options include "bing", "afinn", "loughran", "nrc", "nrc_eil", or "nrc_vad".}

\item{...}{Additional values passed to \code{get_match()}}
}
\value{
A vector or nested list of sentiments for each value of \code{x}.
}
\description{
Get sentiment matches of values in a vector
}
\examples{
my_values <- c("I", "am", "happy")

get_sentiment(my_values)
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_frequency}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()},
\code{\link{is_new}()}
}
\concept{vectorized functions}
