% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_texts.R
\name{identify_by}
\alias{identify_by}
\title{Choose a new doc_id column}
\usage{
identify_by(data, ..., inorder = TRUE, sep = "_")
}
\arguments{
\item{data}{A data frame, potentially with a \code{doc_id} column}

\item{...}{The column or columns that will become the new identifier}

\item{inorder}{Whether to establish \code{doc_id} order as shown in the document}

\item{sep}{Separator between values when identifying by multiple columns}
}
\value{
A data frame with a redefined doc_id column
}
\description{
Choose a new doc_id column
}
\examples{
if (FALSE) {
  corpus |>
    load_texts() |>
    identify_by(author)
}
}
