% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_texts.R
\name{identify_by}
\alias{identify_by}
\title{Choose a new doc_id column}
\usage{
identify_by(data, column, drop = TRUE)
}
\arguments{
\item{data}{A data frame, potentially with a \code{doc_id} column}

\item{column}{The column that will become the new identifier}

\item{drop}{Whether to drop the prior \code{doc_id} column or retain it}
}
\value{
A data frame with a new doc_id column
}
\description{
Choose a new doc_id column
}
\examples{
if (FALSE) {
  corpus |>
    load_texts() |>
    identify_by(author, drop = FALSE)
}
}
