% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_ramp.R
\name{is_new}
\alias{is_new}
\title{Check for new words in a vocabulary}
\usage{
is_new(x)
}
\arguments{
\item{x}{A vector, such as a column of character strings}
}
\value{
A logical vector
}
\description{
Check for new words in a vocabulary
}
\examples{
c("cat", "dog", "dog", "bat", "dog") |>
  is_new()
}
\seealso{
Other vectorized functions: 
\code{\link{get_cumulative_vocabulary}()},
\code{\link{get_frequency}()},
\code{\link{get_hir}()},
\code{\link{get_htr}()},
\code{\link{get_idf_by}()},
\code{\link{get_match}()},
\code{\link{get_sentiment}()},
\code{\link{get_tf_by}()},
\code{\link{get_tfidf_by}()},
\code{\link{get_ttr}()},
\code{\link{is_hapax}()}
}
\concept{vectorized functions}
