% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_titles.R
\name{italicize_titles}
\alias{italicize_titles}
\title{Italicize document titles in a table or figure}
\usage{
italicize_titles(x, ...)
}
\arguments{
\item{x}{A table data object created with \code{tabulize()} or a figure created with \code{visualize()}}

\item{col}{The column to italicize}
}
\value{
A gt table data object or a ggplot2 figure
}
\description{
Italicize document titles in a table or figure
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(gt)
library(tmtyro)

penguins_gt <-
  penguins |>
  select(-year) |>
  summarize(
    across(
      matches("_len$|_dep$"), mean, na.rm = TRUE),
    .by = c(species, island, sex)) |>
  gt() |>
  fmt_number() |>
  tab_spanner(
    "bill",
    columns = starts_with("bill_")) |>
  tab_spanner(
    "flipper",
    starts_with("flip")) |>
  cols_label(
    bill_len = "length",
    bill_dep = "depth",
    flipper_len = "length") |>
  sub_missing()

penguins_gt |>
  collapse_rows(species) |>
  collapse_rows(island) |>
  italicize_titles(species)
}\if{html}{\out{</div>}}
}

\seealso{
Other table helpers: 
\code{\link{collapse_rows}()},
\code{\link{tabulize}()}

Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{table helpers}
\concept{visualizing helpers}
