% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_texts.R
\name{load_texts}
\alias{load_texts}
\title{Load a folder or data frame of texts}
\usage{
load_texts(
  src = "data",
  name = ".txt",
  word = TRUE,
  lemma = FALSE,
  lemma_replace = FALSE,
  to_lower = TRUE,
  remove_names = FALSE,
  pos = FALSE,
  poetry = FALSE,
  paragraph = TRUE,
  n = 1L,
  ...
)
}
\arguments{
\item{src}{Either a string identifying the name of a directory containing texts or a data frame containing an unnested column called "text" and one column with a name ending in "_id". Files should either be stored in a directory within the project folder or under the subdirectory called "data". Defaults to "data" to load texts from that directory.}

\item{name}{What naming pattern to search for in this folder. Defaults to ".txt".}

\item{word}{Whether to split one word per line. Defaults to TRUE.}

\item{lemma}{Whether to lemmatize the text. When \code{word} is TRUE, adds a new column called \code{lemma}. This step can add a lot of time, so it defaults to FALSE.}

\item{lemma_replace}{When \code{lemma} and \code{word} are both TRUE, toggles whether to replace the \code{word} column with the lemmatized tokens. Defaults to FALSE}

\item{to_lower}{When \code{word} is TRUE, toggles whether to convert all words to lowercase. Defaults to TRUE.}

\item{remove_names}{When \code{word} is TRUE, toggles whether to remove words that only appear with the form of initial capitals. Defaults to FALSE.}

\item{pos}{Whether to add a column for part-of-speech tag. This step can add a lot of time, so it defaults to FALSE.}

\item{poetry}{Whether to detect and indicate stanza breaks and line breaks. Defaults to FALSE.}

\item{paragraph}{Whether to detect paragraph breaks for prose. Defaults to TRUE.}

\item{n}{The number of words per row. By default, \code{load_texts()} unnests a text one word at a time using a column called \code{word}. When \code{n} is a value greater than 1, \code{load_texts()} will instead use \code{\link[tidytext:unnest_tokens]{tidytext::unnest_tokens()}} with \code{token = "ngrams"} to create a column called \code{ngram}.}

\item{...}{Additional arguments passed along to \code{\link[tidytext:unnest_tokens]{tidytext::unnest_tokens()}} for use with \code{tokenizers}}
}
\value{
A data frame with two to five columns and one row for each token (optionally, one row for each paragraph or one row for each line) in the corpus.
}
\description{
\code{load_texts()} loads a corpus from a folder of texts or a data frame and prepares it for further study using tidytext principles. By default, \code{load_texts()} will add paragraph numbers (suitable for prose), and unnest at the word level, but options exist to change these defaults for poetry, to avoid unnesting, and even to remove words that seem like proper nouns or to apply techniques of natural language processing for lemmatizing words or tagging their parts of speech.
}
\examples{
\dontrun{
mysteries <-
  load_texts("mystery-novels")

dickinson <-
  load_texts("dickinson-poems",
             poetry = TRUE)

# `load_texts()` can also be used with
# a traditional tidytext workflow:
mysteries <-
  load_texts("mystery-novels",
             word = FALSE,
             to_lower = FALSE) |>
  tidytext::unnest_tokens(word, text)
}
}
