% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{make_dictionary}
\alias{make_dictionary}
\title{Create a lexicon}
\usage{
make_dictionary(definitions, name = NULL)
}
\arguments{
\item{definitions}{A list of named word vectors}

\item{name}{The kind of dictionary}
}
\value{
A data frame with two columns, "word" and \if{html}{\out{<name>}}
}
\description{
\code{make_dictionary()} creates a dictionary for use with \code{add_dictionary()}.
}
\examples{
emoji_weather <- make_dictionary(
  list(
    "ï¸â˜”ï¸" = c("rain", "rains", "rainy", "raining"),
    "ï¸â›ˆï¸" = c("storm", "storms", "stormy", "storming"),
    "â˜ï¸" = c("cloud", "clouds", "cloudy"),
    "ðŸŒž" = c("sun", "sunny"),
    "ðŸŒ«ï¸" = c("fog", "fogs", "foggy", "mist", "misty"),
    "ðŸŒ¬ï¸" = c("wind", "winds", "windy"),
    "ï¸â„ï¸" = c("snow", "snows", "snowing")),
  name = "weather")

border_states <- make_dictionary(
  definitions = list(
    "Canada" = c(
      "Alaska", "Washington", "Idaho",
      "Montana", "North_Dakota", "Minnesota",
      "Wisconsin", "Michigan", "Ohio",
      "Pennsylvania", "New_York", "Vermont",
      "New_Hampshire", "Maine", "AK", "WA",
      "ID", "MT", "ND", "MN", "WI", "MI",
      "OH", "PA", "NY", "VT", "NH", "ME"),
    "Mexico" = c(
      "California", "Arizona", "New_Mexico",
      "Texas", "CA", "AZ", "NM", "TX")),
  name = "borders")
}
