% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg.R
\name{parse_html}
\alias{parse_html}
\title{Read HTML headers and text from file}
\usage{
parse_html(html, title = TRUE)
}
\arguments{
\item{html}{A file in HTML format}

\item{title}{Whether to keep H1 tags even when there is only one unique value}
}
\value{
A data frame with a column called \code{text} and header columns called \code{title}, \code{part}, \code{section}, and \code{subsection} as needed. Header columns are limited to page elements tagged as h1, h2, h3, or h4.
}
\description{
Read HTML headers and text from file
}
\examples{
if (FALSE) {
  ulysses <- parse_html("4300.htm")
}
}
