% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{plot_bigrams}
\alias{plot_bigrams}
\title{Visualize bigram chains}
\usage{
plot_bigrams(
  df,
  feature = word,
  random_seed = TRUE,
  set_seed = NULL,
  edge_color = NULL,
  top_n = 25
)
}
\arguments{
\item{df}{A tidy data frame potentially containing a column called "word" or columns called "word_1" and "word_2".}

\item{feature}{The feature to use when constructing ngrams}

\item{random_seed}{Whether to randomize the creation of the network chart.}

\item{set_seed}{A specific seed to use if not random}

\item{edge_color}{The color of the edges. When the \code{ggraph} package is not loaded, the default \code{edge_color=NULL} setting presents edges as gray. When that package is is loaded, a \code{NULL} edge color colors edges black. Whatever the color defined by this parameter, loading \code{ggraph} enables transparency to show strength of each bigram connection.}

\item{top_n}{The number of pairs to visualize}
}
\value{
A ggplot2 object
}
\description{
Adapted from \verb{[visualize_bigrams()](https://www.tidytextmining.com/ngrams#visualizing-bigrams-in-other-texts)} in Julia Silge and David Robinson's *Text Mining with R.
}
\examples{
\dontrun{
# It isn't necessary to use add_ngrams()
df |>
  plot_bigrams()

# Adding them first allows for filtering steps
df |>
  add_ngrams() |>
  drop_stopwords(word_1) |>
  drop_stopwords(word_2) |>
  plot_bigrams()

# Only bigrams influence the visualization These show the same networks:
df |>
  add_ngrams(2) |>
  plot_bigrams()

df |>
  add_ngrams(4) |>
  plot_bigrams()

}

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  plot_bigrams()

# Loading `ggraph` enables edge to show connection strengths
library(ggraph)

austen |>
  plot_bigrams()

austen |>
  add_ngrams(2) |>
  drop_stopwords(feature = word_1) |>
  drop_stopwords(feature = word_2) |>
  plot_bigrams()

austen |>
  dplyr::filter(doc_id == "northanger") |>
  plot_bigrams(top_n = 80) |>
  change_colors("lightgreen")
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_htr}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{visualize}()}

Other n-gram helpers: 
\code{\link{add_ngrams}()},
\code{\link{combine_ngrams}()},
\code{\link{separate_ngrams}()}
}
\concept{n-gram helpers}
\concept{visualizing helpers}
