% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_document_top_n.R
\name{plot_doc_word_bars}
\alias{plot_doc_word_bars}
\title{Plot bar graphs of frequent features}
\usage{
plot_doc_word_bars(
  data,
  rows = 1:10,
  by = doc_id,
  feature = word,
  inorder = TRUE,
  reorder_y = NULL,
  color_y = FALSE,
  allow_y_blanks = 2,
  percents = TRUE,
  label = NULL,
  label_tweak = 2,
  label_inside = FALSE,
  label_color = NULL,
  na_rm = TRUE
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{rows}{The features to show}

\item{by}{The column used for document grouping, with doc_id as the default}

\item{feature}{The column to measure, as in "word" or "lemma"}

\item{inorder}{Whether to retain the factor order of the "by" column}

\item{reorder_y}{Whether to reorder the Y-values by facet}

\item{color_y}{Whether bars should be filled by Y-values}

\item{allow_y_blanks}{The threshold number of blanks to allow on the Y-axis before dropping all blanks and reprinting values for each facet with a freely scaled Y-axis}

\item{percents}{Whether to display word frequencies as percentage instead of raw counts}

\item{label}{Whether to show the value as a label with each bar}

\item{label_tweak}{The numeric value by which to tweak the label, if shown. For percentages, this value adjusts the decimal-point precision. For raw counts, this value adjusts labels' offset from the bars}

\item{label_inside}{Whether to show the value as a label inside each bar}

\item{label_color}{The color to use for the label text. If no value is chosen, labels will be black when \code{label_inside} is \code{FALSE} and white when \code{TRUE}.}

\item{na_rm}{Whether to drop empty features}
}
\value{
A ggplot object
}
\description{
Plot bar graphs of frequent features
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts(lemma = TRUE) |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  plot_doc_word_bars(rows = 1:4)

dubliners |>
  dplyr::filter(doc_id \%in\% c("The Sisters", "The Dead")) |>
  plot_doc_word_bars(feature = lemma, rows = 1:20)
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
\keyword{internal}
