% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_document_top_n.R
\name{plot_doc_word_bars}
\alias{plot_doc_word_bars}
\title{Plot bar graphs of frequent features}
\usage{
plot_doc_word_bars(
  df,
  rows = 1:10,
  by = doc_id,
  feature = word,
  inorder = TRUE,
  reorder_y = NULL,
  color_y = FALSE,
  percents = TRUE,
  label = FALSE,
  label_tweak = 2,
  label_inside = FALSE,
  na_rm = TRUE
)
}
\arguments{
\item{df}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{rows}{The features to show}

\item{by}{The column used for document grouping, with doc_id as the default}

\item{feature}{The column to measure, as in "word" or "lemma"}

\item{inorder}{Whether to retain the factor order of the "by" column}

\item{reorder_y}{Whether to reorder the Y-values by facet}

\item{color_y}{Whether bars should be filled by Y-values}

\item{percents}{Whether to display word frequencies as percentage instead of raw counts}

\item{label}{Whether to show the value as a label with each bar}

\item{label_tweak}{The numeric value by which to tweak the label, if shown. For percentages, this value adjusts the decimal-point precision. For raw counts, this value adjusts labels' offset from the bars}

\item{label_inside}{Whether to show the value as a label inside each bar}

\item{na_rm}{Whether to drop empty features}
}
\value{
A ggplot object
}
\description{
Plot bar graphs of frequent features
}
\examples{
library(tmtyro)
library(dplyr)

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  plot_doc_word_bars()

austen |>
  mutate(
    pos_pair = paste(pos, lead(pos)),
    `adjective + noun bigram` = paste(word, lead(word))) |>
  filter(stringr::str_detect(pos_pair, "JJ N")) |>
  standardize_titles() |>
  plot_doc_word_bars(
    rows = 1:5,
    feature = `adjective + noun bigram`,
    percents = FALSE,
    label = TRUE,
    label_inside = TRUE,
    label_tweak = -1)
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_htr}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
