% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_document_top_n.R
\name{plot_doc_word_heatmap}
\alias{plot_doc_word_heatmap}
\title{Plot a heatmap of ranked features}
\usage{
plot_doc_word_heatmap(df, num = 10, by = doc_id, feature = word, label = TRUE)
}
\arguments{
\item{df}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{num}{The number of ranks to show, not counting ties}

\item{by}{The column used for document grouping, with doc_id as the default}

\item{feature}{The column to measure, as in "word" or "lemma"}

\item{label}{Whether to show the rank as a label in the heatmap}
}
\value{
A ggplot object
}
\description{
Plot a heatmap of ranked features
}
\examples{
library(tmtyro)
library(dplyr)

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  filter(pos \%in\% c("NN", "NNS")) |>
  plot_doc_word_heatmap()

austen |>
  filter(pos \%in\% c("JJ", "RB")) |>
  plot_doc_word_heatmap(feature = lemma)

}
