% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_document_top_n.R
\name{plot_doc_word_heatmap}
\alias{plot_doc_word_heatmap}
\title{Plot a heatmap of ranked features}
\usage{
plot_doc_word_heatmap(
  .data,
  num = 10,
  by = doc_id,
  feature = word,
  label = TRUE,
  na_color = "white",
  colorset = "viridis",
  line_color = "gray"
)
}
\arguments{
\item{.data}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{num}{The number of ranks to show, not counting ties}

\item{by}{The column used for document grouping, with doc_id as the default}

\item{feature}{The column to measure, as in "word" or "lemma"}

\item{label}{Whether to show the rank as a label in the heatmap}

\item{na_color}{The color to show NA values; default is "white"}

\item{colorset}{The viridis color palette to use; alternatively, a two-item vector of colors to use as gradient}

\item{line_color}{The color of grid lines in the heatmap}
}
\value{
A ggplot object
}
\description{
Plot a heatmap of ranked features
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  dplyr::filter(pos \%in\% c("NN", "NNS")) |>
  plot_doc_word_heatmap()

austen |>
  dplyr::filter(pos \%in\% c("JJ", "RB")) |>
  plot_doc_word_heatmap(feature = lemma, colorset = c("white", "red"))

}
