% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_variety.R
\name{plot_hapax}
\alias{plot_hapax}
\title{Visualize a sample of hapax legomena projected on faceted curves of vocabulary growth over time}
\usage{
plot_hapax(
  df,
  prop = 0.01,
  x = progress_words,
  y = vocabulary,
  by = doc_id,
  descriptive_labels = TRUE,
  feature = hapax
)
}
\arguments{
\item{df}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{prop}{The proportion of hapax to sample. The chart can become illegible with proportions over ~1\%}

\item{x}{The progress column to show. Default option is progress_percent, but progress_words is also appropriate.}

\item{y}{The Y-axis variable to chart. Default value is the cumulative vocabulary size.}

\item{by}{A grouping column, such as doc_id}

\item{descriptive_labels}{A toggle for disabling descriptive labels of progress_percent on the X-axis}

\item{feature}{The column to check for new features. Defaults to \code{hapax}, but the function might also be used with \code{new_word} instead to plot a sample of new additions to documents' vocabularies.}
}
\value{
A ggplot object
}
\description{
Visualize a sample of hapax legomena projected on faceted curves of vocabulary growth over time
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen_measured <- austen |>
  measure_lexical_variety()

austen_measured |>
  standardize_titles() |>
  plot_hapax()
}
