% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_variety.R
\name{plot_hir}
\alias{plot_hir}
\title{Show hapax introduction ratio over time}
\usage{
plot_hir(
  data,
  x = progress,
  by = doc_id,
  identity = doc_id,
  descriptive_labels = TRUE,
  labeling = c("point", "inline", "inset"),
  log_y = TRUE
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing a column called "doc_id" and "word"}

\item{x}{A column showing the cumulative progress of documents}

\item{by}{A grouping column for colors and labels}

\item{identity}{A grouping column for lines}

\item{descriptive_labels}{A toggle for disabling descriptive labels of progress_percent on the X-axis}

\item{labeling}{Options for labeling groups:
\itemize{
\item \code{"point"} labels the final value
\item \code{"inline"} prints the label within a smoothed curve
\item \code{"inset"} prints a legend within the plot area
\item Anything else prints a legend to the right of the plot area.
}}

\item{log_y}{A toggle for logarithmic scaling to the Y-axis; defaults to TRUE}
}
\value{
A ggplot object
}
\description{
Show hapax introduction ratio over time
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners_measured <- dubliners |>
  add_vocabulary()

dubliners_measured |>
  standardize_titles() |>
  plot_hir()
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
\keyword{internal}
