% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{plot_tf_idf}
\alias{plot_tf_idf}
\title{Visualize the top terms by tf-idf}
\usage{
plot_tf_idf(
  data,
  rows = 1:10,
  by = doc_id,
  feature = word,
  label = FALSE,
  label_tweak = 2,
  label_inside = FALSE
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{rows}{The rows of terms to chart in each document}

\item{by}{A column containing document grouping}

\item{feature}{A column containing the terms to be measured across document groupings}

\item{label}{Not yet working}

\item{label_tweak}{Not yet working}

\item{label_inside}{Not yet working}
}
\value{
A ggplot object
}
\description{
Visualize the top terms by tf-idf
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners |>
  plot_tf_idf()
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{italicize_titles}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{tf-idf helpers}
\concept{visualizing helpers}
\keyword{internal}
