% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{plot_tf_idf}
\alias{plot_tf_idf}
\title{Visualize the top terms by tf-idf}
\usage{
plot_tf_idf(
  df,
  num = 10,
  by = doc_id,
  feature = word,
  label = FALSE,
  label_tweak = 2,
  label_inside = FALSE,
  colorset = "default",
  outline_color = NA
)
}
\arguments{
\item{df}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{num}{The number of terms to chart in each document}

\item{by}{A column containing document grouping}

\item{feature}{A column containing the terms to be measured across document groupings}

\item{label}{Not yet working}

\item{label_tweak}{Not yet working}

\item{label_inside}{Not yet working}

\item{colorset}{The color palette to use, whether "default", "okabe", or one of the named qualitative palettes from Viridis or Color Brewer}

\item{outline_color}{The color to use for the outside of each bar. By default, no color is used.}
}
\value{
A ggplot object
}
\description{
Visualize the top terms by tf-idf
}
\examples{
library(tmtyro)
library(dplyr)

austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen |>
  plot_tf_idf()

austen |>
  filter(pos \%in\% c("NN", "NNS"))
  plot_tf_idf(feature = lemma)
}
