% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_variety.R
\name{plot_ttr}
\alias{plot_ttr}
\title{Show type-token ratio over time}
\usage{
plot_ttr(
  df,
  x = progress_words,
  by = doc_id,
  identity = doc_id,
  descriptive_labels = TRUE,
  labeling = c("point", "inline", "axis", "inset"),
  log_y = TRUE
)
}
\arguments{
\item{df}{A tidy data frame, potentially containing a column called "doc_id" and "word"}

\item{x}{The progress column to show. Default option is progress_percent, but progress_words is also appropriate.}

\item{by}{A grouping column for colors and labels}

\item{identity}{A grouping column for lines}

\item{descriptive_labels}{A toggle for disabling descriptive labels of progress_percent on the X-axis}

\item{labeling}{Options for labeling groups:
\itemize{
\item \code{"point"} labels the final value
\item \code{"inline"} prints the label within a smoothed curve
\item \code{"axis"} prints labels where a secondary Y-axis might go
\item \code{"inset"} prints a legend within the plot area
\item Anything else prints a legend to the right of the plot area.
}}

\item{log_y}{A toggle for logarithmic scaling to the Y-axis; defaults to TRUE}
}
\value{
A ggplot object
}
\description{
Show type-token ratio over time
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

austen_measured <- austen |>
  measure_lexical_variety()

austen_measured |>
  standardize_titles() |>
  plot_ttr(labeling = "inline")

austen_measured |>
  standardize_titles() |>
  plot_ttr()
}
