% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_variety.R
\name{plot_vocabulary}
\alias{plot_vocabulary}
\title{Show vocabulary growth}
\usage{
plot_vocabulary(
  data,
  x = progress_words,
  by = doc_id,
  identity = NULL,
  descriptive_labels = TRUE,
  labeling = c("point", "inset", "inline", "axis")
)
}
\arguments{
\item{data}{A tidy data frame, potentially containing columns called "doc_id" and "word"}

\item{x}{A column showing the cumulative count of words}

\item{by}{A grouping column for colors and labels}

\item{identity}{A grouping column for lines}

\item{descriptive_labels}{A toggle for disabling descriptive labels of progress_percent on the X-axis}

\item{labeling}{Options for labeling groups:
\itemize{
\item \code{"point"} labels the final value
\item \code{"inline"} prints the label within a smoothed curve
\item \code{"axis"} prints labels where a secondary Y-axis might go
\item \code{"inset"} prints a legend within the plot area
\item Anything else prints a legend to the right of the plot area.
}}
}
\value{
A ggplot object
}
\description{
\code{plot_vocabulary()} visualizes the vocabulary growth as new words are used in each document.
}
\examples{
\dontrun{
dubliners <- get_gutenberg_corpus(2814) |>
  load_texts() |>
  identify_by(part) |>
  standardize_titles()

dubliners_measured <- dubliners |>
  add_vocabulary()

dubliners_measured |>
  plot_vocabulary(progress_percent)

dubliners_measured |>
  plot_vocabulary()

  get_micusp_corpus(
    discipline \%in\% c("Physics", "Economics")) |>
    load_texts() |>
    add_vocabulary() |>
    plot_vocabulary(by = discipline)
}
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_hir}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{theme_tmtyro}()},
\code{\link{visualize}()}
}
\concept{visualizing helpers}
\keyword{internal}
