% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_titles.R
\name{standardize_titles}
\alias{standardize_titles}
\title{Standardize document titles}
\usage{
standardize_titles(.data, title = doc_id)
}
\arguments{
\item{.data}{A tidy data frame, potentially containing a title column called "doc_id". Alternatively, a simple character vector of titles.}

\item{title}{A column containing the titles to be standardized}
}
\value{
A data frame with one column adjusted. If .data is a character vector instead of a data frame, then a character vector is returned.
}
\description{
Useful especially for visualizations. \code{standardize_titles} applies some English-language conventions, including converting underscores to spaces, capitalizing important words, removing leading articles, and dropping subtitles.
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

##### Standardizing strings #####
# Before `standardize_titles()`
unique(austen$doc_id)

# After `standardize_titles()`
unique(austen$doc_id) |>
  standardize_titles()

##### Standardizing a data frame #####

austen_measured <- austen |>
  add_lexical_variety()

# Before `standardize_titles()`
austen_measured |>
  plot_vocabulary(labeling = "inline")

# After `standardize_titles()`
austen_measured |>
  standardize_titles() |>
  plot_vocabulary(labeling = "inline")
}
