% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.ngrams}
\alias{tabulize.ngrams}
\title{Prepare a table of n-gram frequencies}
\usage{
\method{tabulize}{ngrams}(.data, ...)
}
\arguments{
\item{.data}{data processed with one or more functions from \code{tmtyro}}

\item{...}{optional parameters passed along to methods}
}
\description{
Prepare a table of n-gram frequencies
}
\keyword{internal}
