% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize.tf_idf}
\alias{tabulize.tf_idf}
\title{Prepare a table of term frequency--inverse document frequency}
\usage{
\method{tabulize}{tf_idf}(.data, rows = NULL, digits = 5, feature = word, ...)
}
\arguments{
\item{.data}{data processed with one or more functions from \code{tmtyro}}

\item{rows}{Chooses rows to be shown}

\item{digits}{The number of digits to show past the decimal point}

\item{...}{optional parameters passed along to methods}
}
\description{
Prepare a table of term frequency--inverse document frequency
}
\keyword{internal}
