% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_without_caps.R
\name{unnest_without_caps}
\alias{unnest_without_caps}
\title{Split text into words and drop proper nouns}
\usage{
unnest_without_caps(df, output = "word", input = "text", to_lower = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{output}{Output column to be created.}

\item{input}{Input column that gets split by word.}

\item{to_lower}{Whether to convert final words to lowercase.}
}
\value{
A data frame
}
\description{
Split a column of text using \code{\link[tidytext:unnest_tokens]{tidytext::unnest_tokens()}}, flattening the table into one token per row while also omitting any token that is present only in a capitalized form.
}
\examples{
\dontrun{
mysteries <-
  load_texts("mystery-novels",
             to_lower = FALSE) |>
  unnest_without_caps()

# Since `unnest_without_caps()` is
# incorporated into `load_texts()`,
# it may be unnecessary for many
# scenarios.
mysteries <-
  load_texts("mystery-novels",
             remove_names = TRUE)
  }
}
