% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize output}
\usage{
visualize(.data, ..., type = NULL)
}
\arguments{
\item{.data}{data processed with one or more functions from \code{tmtyro}}

\item{...}{optional parameters}

\item{type}{visualization type preferred, used after \code{\link[=add_lexical_variety]{add_lexical_variety()}} and (eventually) \code{\link[=make_topic_model]{make_topic_model()}}}
}
\value{
a ggplot2 object
}
\description{
Visualize output
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

# A data frame with `doc_id` and `word` columns will visualize frequency by default
austen |>
   visualize()

# Applying `tmtyro` functions will choose an appropriate visualization
austen |>
   add_sentiment() |>
   visualize()

austen |>
   add_ngrams() |>
   visualize()

austen |>
   add_ngrams() |>
   combine_ngrams() |>
   visualize()

austen |>
   summarize_tf_idf() |>
   visualize()

austen |>
   add_lexical_variety() |>
   visualize()

# Some visualizations are specified with the `type` argument
austen |>
   add_lexical_variety() |>
   visualize(type = "ttr")

austen |>
   add_lexical_variety() |>
   visualize(type = "hapax")

# Other arguments get passed along
austen |>
   add_ngrams() |>
   visualize(node_color = "yellow", top_n = 50)

austen |>
   add_lexical_variety() |>
   visualize()
}
