% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize output}
\usage{
visualize(.data, ...)
}
\arguments{
\item{.data}{data processed with one or more functions from \code{tmtyro}}

\item{...}{
  Arguments passed on to \code{\link[=plot_doc_word_bars]{plot_doc_word_bars}}, \code{\link[=plot_bigrams]{plot_bigrams}}, \code{\link[=plot_vocabulary]{plot_vocabulary}}, \code{\link[=plot_ttr]{plot_ttr}}, \code{\link[=plot_htr]{plot_htr}}, \code{\link[=plot_topic_distributions]{plot_topic_distributions}}, \code{\link[=plot_topic_bars]{plot_topic_bars}}, \code{\link[=plot_topic_wordcloud]{plot_topic_wordcloud}}
  \describe{
    \item{\code{rows}}{The features to show}
    \item{\code{by}}{The column used for document grouping, with doc_id as the default}
    \item{\code{feature}}{The column to measure, as in "word" or "lemma"}
    \item{\code{inorder}}{Whether to retain the factor order of the "by" column}
    \item{\code{reorder_y}}{Whether to reorder the Y-values by facet}
    \item{\code{color_y}}{Whether bars should be filled by Y-values}
    \item{\code{percents}}{Whether to display word frequencies as percentage instead of raw counts}
    \item{\code{label}}{Whether to show the value as a label with each bar}
    \item{\code{label_tweak}}{The numeric value by which to tweak the label, if shown. For percentages, this value adjusts the decimal-point precision. For raw counts, this value adjusts labels' offset from the bars}
    \item{\code{label_inside}}{Whether to show the value as a label inside each bar}
    \item{\code{na_rm}}{Whether to drop empty features}
    \item{\code{random_seed}}{Whether to randomize the creation of the network chart.}
    \item{\code{set_seed}}{A specific seed to use if not random}
    \item{\code{edge_color}}{The color of the edges. When the \code{ggraph} package is not loaded, the default \code{edge_color=NULL} setting presents edges as gray. When that package is is loaded, a \code{NULL} edge color colors edges black. Whatever the color defined by this parameter, loading \code{ggraph} enables transparency to show strength of each bigram connection.}
    \item{\code{top_n}}{The number of pairs to visualize}
    \item{\code{identity}}{A grouping column for lines}
    \item{\code{descriptive_labels}}{A toggle for disabling descriptive labels of progress_percent on the X-axis}
    \item{\code{labeling}}{Options for labeling groups:
\itemize{
\item \code{"point"} labels the final value
\item \code{"inline"} prints the label within a smoothed curve
\item \code{"axis"} prints labels where a secondary Y-axis might go
\item \code{"inset"} prints a legend within the plot area
\item Anything else prints a legend to the right of the plot area.
}}
    \item{\code{log_y}}{A toggle for logarithmic scaling to the Y-axis; defaults to TRUE}
    \item{\code{topics}}{The topic numbers to view}
  }}
}
\value{
a ggplot2 object
}
\description{
\code{visualize()} provides a simple method for displaying results. Based on previous functions used, \code{visualize()} will choose a method, resolving to one of the other visualizing helpers.
}
\note{
For some visualizations, an optional \code{type} parameter may be helpful to change the visualization. For example, setting \code{type = "htr"}, \code{type = "ttr"}, or \code{type = "hapax"} after \code{\link[=add_vocabulary]{add_vocabulary()}} will emphasize different columns added by that function. Similarly, \code{type = "cloud"} or \code{type = "wordcloud"} will show topic word clouds after \code{\link[=make_topic_model]{make_topic_model()}}, and \code{type = "heatmap"} will show an alternative visualization for word frequencies.
}
\examples{
austen <- "austen.rds" |>
  system.file(package = "tmtyro") |>
  readRDS()

# A data frame with `doc_id` and `word` columns will visualize frequency by default
austen |>
   visualize()

# Applying `tmtyro` functions will choose an appropriate visualization

austen |>
   add_ngrams() |>
   visualize()

austen |>
   add_ngrams() |>
   combine_ngrams() |>
   visualize()

austen |>
   summarize_tf_idf() |>
   visualize()

austen |>
   add_vocabulary() |>
   visualize()

if (FALSE) { # sentiment requires interaction on first load
  austen |>
     add_sentiment() |>
     visualize()
}

# Some visualizations are specified with the `type` argument
austen |>
   add_vocabulary() |>
   visualize(type = "ttr")

austen |>
   add_vocabulary() |>
   visualize(type = "hapax")

# Other arguments get passed along
austen |>
   add_ngrams() |>
   visualize(top_n = 25)

austen |>
   add_vocabulary() |>
   visualize(labeling = "inline")
}
\seealso{
Other visualizing helpers: 
\code{\link{change_colors}()},
\code{\link{plot_bigrams}()},
\code{\link{plot_doc_word_bars}()},
\code{\link{plot_doc_word_heatmap}()},
\code{\link{plot_hapax}()},
\code{\link{plot_htr}()},
\code{\link{plot_tf_idf}()},
\code{\link{plot_topic_bars}()},
\code{\link{plot_topic_distributions}()},
\code{\link{plot_topic_wordcloud}()},
\code{\link{plot_ttr}()},
\code{\link{plot_vocabulary}()}
}
\concept{visualizing helpers}
